+++ modules/core/src/passes/layers-pass.ts
@@ -95,6 +95,8 @@ export default class LayersPass extends Pass {
       return this._drawLayers(renderPass, options);
     } finally {
       renderPass.end();
+      // TODO(ibgreen): WebGPU - submit may not be needed here but initial port had issues with out of render loop rendering
+      this.device.submit();
     }
   }